reader/requirements.txt
nats-py
pytest